refactor: ynet.FrameConn -> fame.Conn
