+++ core/server.go
@@ -13,14 +13,13 @@ import (
 	"github.com/yomorun/yomo/core/auth"
 	"github.com/yomorun/yomo/core/frame"
 	"github.com/yomorun/yomo/core/metadata"
-	ynet "github.com/yomorun/yomo/core/net"
 	"github.com/yomorun/yomo/core/router"
 	"golang.org/x/exp/slog"
 
 	// authentication implements, Currently, only token authentication is implemented
 	_ "github.com/yomorun/yomo/pkg/auth"
 	"github.com/yomorun/yomo/pkg/frame-codec/y3codec"
-	yquic "github.com/yomorun/yomo/pkg/netimpl/quic"
+	yquic "github.com/yomorun/yomo/pkg/listener/quic"
 	pkgtls "github.com/yomorun/yomo/pkg/tls"
 	oteltrace "go.opentelemetry.io/otel/trace"
 )
@@ -57,7 +56,7 @@ type Server struct {
 	opts               *serverOptions
 	frameHandler       FrameHandler
 	connHandler        ConnHandler
-	listener           ynet.Listener
+	listener           frame.Listener
 	logger             *slog.Logger
 	tracerProvider     oteltrace.TracerProvider
 }
@@ -153,7 +152,7 @@ func (s *Server) Serve(ctx context.Context, conn net.PacketConn) error {
 	}
 }
 
-func (s *Server) handshake(fconn ynet.FrameConn) (bool, router.Route, *Connection) {
+func (s *Server) handshake(fconn frame.Conn) (bool, router.Route, *Connection) {
 	var gerr error
 
 	defer func() {
@@ -215,7 +214,7 @@ func (s *Server) handleConnRoute(conn *Connection, route router.Route) {
 	}
 }
 
-func (s *Server) handleFrameConn(fconn ynet.FrameConn, logger *slog.Logger) {
+func (s *Server) handleFrameConn(fconn frame.Conn, logger *slog.Logger) {
 	ok, route, conn := s.handshake(fconn)
 	if !ok {
 		logger.Error("handshake failed")
@@ -230,7 +229,7 @@ func (s *Server) handleFrameConn(fconn ynet.FrameConn, logger *slog.Logger) {
 	_ = s.connector.Remove(conn.ID())
 }
 
-func (s *Server) handleHandshakeFrame(fconn ynet.FrameConn, hf *frame.HandshakeFrame) (*Connection, error) {
+func (s *Server) handleHandshakeFrame(fconn frame.Conn, hf *frame.HandshakeFrame) (*Connection, error) {
 	md, ok := auth.Authenticate(s.opts.auths, hf)
 
 	if !ok {
@@ -390,7 +389,7 @@ func (s *Server) dispatchToDownstreams(c *Context) error {
 	return nil
 }
 
-func closeServer(downstreams map[string]Downstream, connector *Connector, listener ynet.Listener, router router.Router) error {
+func closeServer(downstreams map[string]Downstream, connector *Connector, listener frame.Listener, router router.Router) error {
 	for _, ds := range downstreams {
 		ds.Close()
 	}
